requirements.txt
# Python dependencies for the project

PyGithub
requests
pandas
